This is the unity project with the player character and the animations
Animations are in FigtherMotionsVolume1.
Character is in Models/Characters.

There are also a few scripts and animation controllers.
Shared_Controller lets characters play a single animation when a key is pressed (via a script).

BlueGuy_Animator just plays all the animations in the FigtherMotionsVolume1 pack one after another.

I also started making the player 1 controller, I put most of the animations that we should used in the controller but it is not finished yet. It also needs a script to let the player control the character.